Introduce a low-level `protocol` module
This is quite a large refactor of the protocol-related items in an
attempt to make a few improvements, mostly wrt to keeping in mind future
plans for both client and server implementations:

- Allow for more granular `read`ing and `write`ing of protocol types.
- Allow for `write`ing protocol types using any type that implements
  `std::io::Write`. Useful for re-using slices of bytes.
- Follow the conventions set by the `byteorder` crate dependency.
- Provide relevant protocol documentation along with each protocol type.
- Provide a more uniform interface between reading and writing,
  hopefully making things a bit easier for future server implementation.

Most of these changes are reflections of lessons I've learned
implementing other protocols in the past including
[ether-dream](https://github.com/nannou-org/ether-dream),
[citp](https://github.com/nannou-org/citp),
[m3u](https://github.com/mitchmindtree/m3u) and other commercial work.
Below I'll give some more details on these changes in the hope that it
provides a bit more clarity and makes the changes a bit easier to
navigate!

The `protocol` module
---------------------

The `protocol` module contains all types and constants defined in the
NTP RFC and provides method implementations that `write` and `read`
bytes in network endian order in a manner that precisely matches the
specification. I'm envisioning the `protocol` module as a common,
low-level layer that may be shared between `client` and `server`
implementations.

All protocol types have been consolidated within this `protocol` module.
I find this to make it a little easier to identify types that represent
the low-level protocol description, first giving an overview of the
model (the types) and then their behaviour (inherent implementations
followed by read/write implementations).

`ReadBytes` and `WriteBytes` traits
-----------------------------------

These traits can be thought of as extensions to the `byteorder`
`ReadBytesExt` and `WriteBytesExt` traits respectively that are specific
to NTP and allow for reading and writing NTP types to and from
network-endian octets.

In the same way that the byteorder traits enable `io::Read` and `Write`
types to read and write the various primitive types, the ntp `ReadBytes`
and `WriteBytes` traits enable those same readers and writers to read
and write ntp protocol types such as `ShortFormat`, `TimestampFormat`,
`Packet`, and so on.

Documentation
-------------

I've attempted to improve documentation for the protocol module items by
adapting (and in some cases directly copying) the specification as
described in the NTP RFC. I've attempted to format the doc comments in a
manner that should make them easy to peruse and navigate from the module
level docs.

Example
-------

The example has been updated, though as the API is pretty much the same
not much has changed. I've slightly modified it to print each of the
given timestamps in local time as this personally helped me to intuit
how the NTP packets worked a little better.

Tests
-----

The tests have been moved into a `tests/protocol.rs` module and updated
for the protocol refactor, using the original data and still passing.

Future Work
-----------

Following this, I imagine a nice next step might be to move the
`request` API into a `client` module in an attempt to prepare for the
addition of a `server` module. I still need to read through the protocol
specification more thoroughly before I'll feel comfortable diving any
deeper, however I'm currently in south-east Asia purely for the purpose
of R&D so I'm sure I'll get around to it soon :)

@JeffBelgum Sorry for the large PR! I realise this is a lot to digest,
Feel free to take your time reviewing and ask any question is you're
unsure about any of the changes, etc.