Rename `SizeBytes` to `ConstPackedSizeBytes` to reduce ambiguity
I realised that it might be easy to interpret the purpose of the trait
as a representation of the type's actual size, whereas in reality the
purpose is to represent the size when the type is written to or read
from octets.

Also, I've found it useful to distinguish between types whose packed
size are known at compile time, and those who are dynamic (e.g. types
that may contain variable length lists). Thus, I also added the `Const`
prefix.